Clean up all DB cod, file/folder structure
