+++ db/riak/seeding.js
